+++ footer.php
@@ -47,4 +47,3 @@ if ( ! defined( 'ABSPATH' ) ) {
 
 </body>
 </html>
-